Restructure the project
1. Changes the project structure and puts all source files under
directories in src.
2. Makes the Makefile a bit nicer and easier to navigate.
3. Refactors test_ubi into 3 different files to make it more manageable.
