+++ Makefile
@@ -1,10 +1,8 @@
 # Source files and object files
 SRC_DIR := src
-APP_DIR := app
-TEST_DIR := test
-OBJ_DIR := build/obj
-BIN_DIR := build/bin
-TEST_BDEVS := --bdev ubi0 --bdev ubi_nosync --bdev ubi_directio --bdev ubi_copy_on_read
+APP_DIR := $(SRC_DIR)/app
+LIB_DIR := $(SRC_DIR)/lib
+BIN_DIR := bin
 
 ifneq ($(MAKECMDGOALS),format)
 PKG_CONFIG_PATH = $(SPDK_PATH)/lib/pkgconfig
@@ -13,73 +11,36 @@ SYS_LIB := $(shell PKG_CONFIG_PATH="$(PKG_CONFIG_PATH)" pkg-config --libs --stat
 
 # Above pkg-config command adds two instances of "-lrte_net", which results in
 # bunch of multiple definition errors. Make sure it's not repeated.
-SPDK_DPDK_LIB := $(filter-out -lrte_net,$(SPDK_DPDK_LIB))
-SPDK_DPDK_LIB += -lrte_net
+SPDK_DPDK_LIB := $(filter-out -lrte_net,$(SPDK_DPDK_LIB)) -lrte_net
+endif
 
-# Compiler and linker flags
 CFLAGS := -D_GNU_SOURCE -Iinclude -Wall -g -O3 -I$(SPDK_PATH)/include
 LDFLAGS := -Wl,--whole-archive,-Bstatic $(SPDK_DPDK_LIB) -Wl,--no-whole-archive -luring -Wl,-Bdynamic $(SYS_LIB)
-endif
 
 ifeq ($(COVERAGE),true)
     CFLAGS += -fprofile-arcs -ftest-coverage
 endif
 
-# Automatically generate a list of source files (.c) and object files (.o)
-SRCS := $(wildcard $(SRC_DIR)/*.c)
-OBJS := $(SRCS:$(SRC_DIR)/%.c=$(OBJ_DIR)/%.o)
+LIB_SRCS := $(shell find $(LIB_DIR) -name '*.c')
+LIB_OBJS := $(LIB_SRCS:%.c=%.o)
 
-all: $(BIN_DIR)/vhost_ubi $(BIN_DIR)/test_ubi $(BIN_DIR)/memcheck_ubi $(BIN_DIR)/test_image.raw $(BIN_DIR)/test_disk.raw
-
-$(BIN_DIR)/vhost_ubi: $(OBJS) $(OBJ_DIR)/vhost_ubi.o
-	@mkdir -p $(BIN_DIR)
-	$(CC) $(CFLAGS) -o $@ $^ $(LDFLAGS)
-
-$(BIN_DIR)/test_ubi: $(OBJS) $(OBJ_DIR)/test_ubi.o
-	@mkdir -p $(BIN_DIR)
-	$(CC) $(CFLAGS) -o $@ $^ $(LDFLAGS)
-
-$(BIN_DIR)/memcheck_ubi: $(OBJS) $(OBJ_DIR)/memcheck_ubi.o
-	@mkdir -p $(BIN_DIR)
-	$(CC) $(CFLAGS) -o $@ $^ $(LDFLAGS)
-
-$(OBJ_DIR)/%.o: $(SRC_DIR)/%.c
-	@mkdir -p $(OBJ_DIR)
-	$(CC) $(CFLAGS) -c $< -o $@
-
-$(OBJ_DIR)/%.o: $(APP_DIR)/%.c
-	@mkdir -p $(OBJ_DIR)
-	$(CC) $(CFLAGS) -c $< -o $@
-
-$(OBJ_DIR)/%.o: $(TEST_DIR)/%.c
-	@mkdir -p $(OBJ_DIR)
-	$(CC) $(CFLAGS) -c $< -o $@
-
-$(BIN_DIR)/test_image.raw:
-	dd if=/dev/random of=$@ bs=1048576 count=40
-
-$(BIN_DIR)/test_disk.raw: $(BIN_DIR)/test_image.raw
-	cp $< $@
-	truncate --size 100M $@
-
-check:
-	sudo ./build/bin/test_ubi --cpumask [0,1,2] --json test/test_conf.json \
-		--json-ignore-init-errors $(TEST_BDEVS)
-
-valgrind:
-	sudo valgrind ./build/bin/memcheck_ubi --cpumask [0] \
-		--json-ignore-init-errors --json test/test_conf.json $(TEST_BDEVS)
-
-coverage:
-	lcov --capture --directory . --exclude=`pwd`/test/'*.c' --no-external --output-file coverage.info > /dev/null
-	genhtml coverage.info --output-directory coverage_report
-
-# Clean up build artifacts
-clean:
-	@rm -rf $(OBJ_DIR) $(BIN_DIR) coverage.info coverage_report
-
-# Automatically format source files
-format:
-	find . -regex '.*\.\(c\|h\)$$' -exec clang-format -i {} \;
+APP_TARGETS = $(BIN_DIR)/vhost_ubi
 
 .PHONY: all clean
+all: # forward declaration
+
+include src/test/build.mk
+
+all: $(APP_TARGETS) $(TEST_TARGETS)
+
+$(BIN_DIR)/%: $(APP_DIR)/%.c $(LIB_OBJS)
+	$(info Building $@ ...)
+	@mkdir -p $(@D)
+	@$(CC) $(CFLAGS) $< $(LIB_OBJS) -o $@ $(LDFLAGS)
+
+COVERAGE_FILES := $(shell find $(LIB_DIR) -name '*.gc*')
+clean:
+	@rm -rf $(LIB_OBJS) $(BIN_DIR) coverage.info coverage_report $(COVERAGE_FILES)
+
+format:
+	find . -regex '.*\.\(c\|h\)$$' -exec clang-format -i {} \;
