+++ src/lib/bdev_ubi_stripe.c
